feat(WR): add alternative pselect support
Tokio AsyncFD doesn't work well on platform like Darwin. Causing
Segmentation fault related to this_binding->unwind.arg being null.
