+++ rust_src/crates/webrender/Cargo.in
@@ -26,11 +26,14 @@ tokio = { version = "1.23", features = ["rt-multi-thread", "sync", "net", "macro
 futures = "0.3.25"
 winit = { version = "0.27.5", default-features = false }
 fontdb = "0.12"
-nix = "0.26"
-tracing = "0.1"
 errno = "0.2"
-polling = "2.5"
-easy-parallel = "3.2"
+spin_sleep = { version = "1.1", optional = true }
+
+[dependencies.nix]
+optional = true
+version = "0.26"
+default-features = false
+features = ["signal"]
 
 [dependencies.webrender]
 git = "https://github.com/servo/webrender.git"
@@ -61,3 +64,4 @@ x11 = ["copypasta/x11", "surfman/sm-x11", "winit/x11"]
 wayland = ["copypasta/wayland", "winit/wayland"]
 capture=["webrender/capture", "webrender/serialize_program"]
 sw_compositor=["webrender/sw_compositor"]
+pselect=["dep:spin_sleep", "dep:nix"]
