+++ rust_src/crates/webrender/src/select/plain.rs
@@ -1,0 +1,89 @@
+use errno::{set_errno, Errno};
+use nix::sys::signal::{self, Signal};
+use std::{
+    cell::RefCell,
+    time::{Duration, Instant},
+};
+
+use libc::{fd_set, sigset_t, timespec};
+use winit::{
+    event::{Event, WindowEvent},
+    platform::run_return::EventLoopExtRunReturn,
+};
+
+use emacs::bindings::make_timespec;
+
+use crate::event_loop::EVENT_BUFFER;
+use crate::event_loop::EVENT_LOOP;
+
+pub fn handle_select(
+    nfds: i32,
+    readfds: *mut fd_set,
+    writefds: *mut fd_set,
+    _exceptfds: *mut fd_set,
+    timeout: *mut timespec,
+    _sigmask: *mut sigset_t,
+) -> i32 {
+    let mut event_loop = EVENT_LOOP.lock().unwrap();
+
+    let deadline = Instant::now()
+        + unsafe { Duration::new((*timeout).tv_sec as u64, (*timeout).tv_nsec as u32) };
+
+    let nfds_result = RefCell::new(0);
+
+    // We mush run winit in main thread, because the macOS platfrom limitation.
+    event_loop.el.run_return(|e, _, control_flow| {
+        control_flow.set_wait_until(deadline);
+
+        match e {
+            Event::WindowEvent { ref event, .. } => match event {
+                WindowEvent::Resized(_)
+                | WindowEvent::KeyboardInput { .. }
+                | WindowEvent::ReceivedCharacter(_)
+                | WindowEvent::ModifiersChanged(_)
+                | WindowEvent::MouseInput { .. }
+                | WindowEvent::CursorMoved { .. }
+                | WindowEvent::Focused(_)
+                | WindowEvent::MouseWheel { .. }
+                | WindowEvent::CloseRequested => {
+                    EVENT_BUFFER.lock().unwrap().push(e.to_static().unwrap());
+
+                    // notify emacs's code that a keyboard event arrived.
+                    match signal::raise(Signal::SIGIO) {
+                        Ok(_) => {}
+                        Err(err) => log::error!("sigio err: {err:?}"),
+                    };
+                    /* Pretend that `select' is interrupted by a signal.  */
+                    set_errno(Errno(libc::EINTR));
+                    debug_assert_eq!(nix::errno::errno(), libc::EINTR);
+                    nfds_result.replace(-1);
+                    control_flow.set_exit();
+                }
+                _ => {}
+            },
+            Event::UserEvent(nfds) => {
+                nfds_result.replace(nfds);
+                control_flow.set_exit();
+            }
+            Event::RedrawEventsCleared => {
+                control_flow.set_exit();
+            }
+            _ => {}
+        };
+    });
+    let ret = nfds_result.into_inner();
+    if ret == 0 {
+        let timespec = unsafe { make_timespec(0, 0) };
+        // Add some delay here avoding high cpu usage on macOS
+        #[cfg(target_os = "macos")]
+        spin_sleep::sleep(Duration::from_millis(16));
+        let nfds =
+            unsafe { libc::pselect(nfds, readfds, writefds, _exceptfds, &timespec, _sigmask) };
+        log::trace!("pselect: {nfds:?}");
+        return nfds;
+    }
+
+    log::trace!("winit event run_return: {ret:?}");
+
+    ret
+}
