+++ src/lisp.h
@@ -3984,7 +3984,7 @@ extern void syms_of_data (void);
 extern void swap_in_global_binding (struct Lisp_Symbol *);
 
 /* Defined in emacs.c */
-extern int main1 (int, char **);
+extern _Noreturn int main1 (int, char **);
 
 /* Defined in cmds.c */
 extern void syms_of_cmds (void);
